# swagger.yml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Library Management System
  description: Library Management System API
tags:
  - name: Book
  - name: Borrow
  - name: Borrower
  
paths:

## Book API ##
  /book/all:
    get:
      summary: Get all books
      tags:
        - Book
      responses:
        '200':
          description: Successful response
        '500':
          description: Internal server error

  /book/:
    post:
      summary: Add a new book
      tags:
        - Book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                isbn:
                  type: string
                title:
                  type: string
                author:
                  type: string
                quantity:
                  type: integer
                shelfNumber:
                  type: integer
              example:
                isbn: '9783161484100'
                title: Book title
                author: Book author
                quantity: 1
                shelfNumber: 0
      responses:
        '201':
          description: Created successfully
        '500':
          description: Internal server error

    patch:
      summary: Update an existing book
      tags:
        - Book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                isbn:
                  type: string
                title:
                  type: string
                author:
                  type: string
                quantity:
                  type: integer
                shelfNumber:
                  type: integer
              example:
                isbn: '9783161484100'
                title: Book title
                author: Book author
                quantity: 1
                shelfNumber: 0
      responses:
        '200':
          description: Updated successfully
        '500':
          description: Internal server error

  /book/{isbn}:
    get:
      summary: Get Book info
      tags:
        - Book
      parameters:
        - in: path
          name: isbn
          required: true
          schema:
            type: string
          description: Book ISBN
      responses:
        '200':
          description: Successful response
        '500':
          description: Internal server error

    delete:
      summary: Delete a book
      tags:
        - Book
      parameters:
        - in: path
          name: isbn
          required: true
          schema:
            type: string
          description: Book ISBN
      responses:
        '200':
          description: Deleted successfully
        '500':
          description: Internal server error

  /book/search/ISBN/{key}:
    get:
      summary: Search existing books by ISBN
      tags:
        - Book
      parameters:
        - in: path
          name: key
          required: true
          schema:
            type: string
          description: Search key
      responses:
        '200':
          description: Successful response
        '500':
          description: Internal server error

  /book/search/Author/{key}:
    get:
      summary: Search existing books by Author
      tags:
        - Book
      parameters:
        - in: path
          name: key
          required: true
          schema:
            type: string
          description: Search key
      responses:
        '200':
          description: Successful response
        '500':
          description: Internal server error

  /book/search/Title/{key}:
    get:
      summary: Search existing books by Title
      tags:
        - Book
      parameters:
        - in: path
          name: key
          required: true
          schema:
            type: string
          description: Search key
      responses:
        '200':
          description: Successful response
        '500':
          description: Internal server error


## Borrow API ##
  /borrow/borrow/:
    post:
      summary: Borrow a book
      tags:
        - Borrow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                isbn:
                  type: string
                email:
                  type: string
                quantity:
                  type: integer
                dueDate:
                  type: date
              example:
                isbn: '9783161484100'
                email: n@n.com
                quantity: 1
                dueDate: 2024-01-01
      responses:
        '200':
          description: Successful response
        '500':
          description: Internal server error

  
  /borrow/return/:
    post:
      summary: Return a borrowed book
      tags:
        - Borrow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                isbn:
                  type: string
                email:
                  type: string
              example:
                isbn: '9783161484100'
                email: n@n.com
      responses:
        '200':
          description: Successful response
        '500':
          description: Internal server error

  
  /borrow/mybooks/{email}:
    get:
      summary: Get all borrowed books of a borrower
      tags:
        - Borrow
      parameters:
        - in: path
          name: email
          required: true
          schema:
            type: string
          description: Borrower email
      responses:
        '200':
          description: Successful response
        '500':
          description: Internal server error


  /borrow/overdue/:
    get:
      summary: Get all overdue borrows
      tags:
        - Borrow
      responses:
        '200':
          description: Successful response
        '500':
          description: Internal server error

## Borrower API ##
  /borrower/:
    post:
      summary: Register a new borrower
      tags:
        - Borrower
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                name:
                  type: string
              example:
                email: n@n.com
                name: John Doe
      responses:
        '201':
          description: Created successfully
        '500':
          description: Internal server error

    patch:
      summary: Update an existing borrower
      tags:
        - Borrower
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                name:
                  type: string
              example:
                email: n@n.com
                name: John Doe
      responses:
        '200':
          description: Updated successfully
        '500':
          description: Internal server error

  /borrower/{email}:
    get:
      summary: Get Borrower info
      tags:
        - Borrower
      parameters:
        - in: path
          name: email
          required: true
          schema:
            type: string
          description: Borrower email
      responses:
        '200':
          description: Successful response
        '500':
          description: Internal server error

    delete:
      summary: Delete borrower
      tags:
        - Borrower
      parameters:
        - in: path
          name: email
          required: true
          schema:
            type: string
          description: Borrower email
      responses:
        '200':
          description: Deleted successfully
        '500':
          description: Internal server error

  /borrower/all:
    get:
      summary: Get all borrowers
      tags:
        - Borrower
      responses:
        '200':
          description: Successful response
        '500':
          description: Internal server error